Renames v2 transformers file
